Reapply code to save previous UNUserNotificationCenterDelegate
This code fixes an issue with foreground notifications in iOS. Without this fix, the on notification callback is called multiple times.

This fix was done originally in c38cf81002d7140c37ffa5f7c35c5ca06e7d61fe, but the file were the fix was applied was deleted in the original repository. Instead of moving the code in the merge commit, I preferred adding a new commit to make it clearer that this code has been moved.